+++ src/server/routes/users.js
@@ -2,6 +2,7 @@ import express from "express";
 import {isValidEmail, isValidAssignedName, isValidFamilyName} from "../validation.mjs";
 import User from "../models/User.mjs";
 import HttpStatusCodes from "../HttpStatusCodes.mjs"
+import { hashPassword, comparePassword } from "../crypto.mjs";
 
 
 export function createUsersRouter({ userRepo }) {
@@ -26,11 +27,11 @@ export function createUsersRouter({ userRepo }) {
             return res.status(HttpStatusCodes.BAD_REQUEST).json({ error: "Error: invalid family name!"});
         }
         
-        //aggiungi altre verifiche
-
+        //hash the password
+        const hashedPassword = await hashPassword(password);
 
         //create new user object
-        const newUser = new User(null, email, assignedName, familyName, password);
+        const newUser = new User(null, email, assignedName, familyName, hashedPassword);
         try {
             const res_ = await userRepo.createUser(newUser);
             return res.status(HttpStatusCodes.CREATED).json(newUser);
@@ -43,7 +44,12 @@ export function createUsersRouter({ userRepo }) {
     router.put("/:id", async (req, res) => {
         //convert id to number
         const id = parseInt(req.params.id);
-        const { email, assignedName, familyName, password} = req.body;
+
+        if (isNaN(id)){ 
+            return res.status(HttpStatusCodes.BAD_REQUEST).json({ error: "Error: id is not a number!" });
+        }
+
+        const { email, assignedName, familyName, password, newPassword} = req.body;
 
         //verify email is valid
         if (!isValidEmail(email)) {
@@ -61,7 +67,38 @@ export function createUsersRouter({ userRepo }) {
         }
 
 
-        //aggiungi altre verifiche
+        //for password changes:
+        //if newPassword is not null -> the user wanna change the password
+        //in this case verify if the old password is correct
+        //fetch the old user from the db anc compare the hashed passwords
+        
+        if (newPassword) {
+            try {
+                const fecthedUser = await userRepo.getUserById(id);
+
+                if (fecthedUser) {
+                    //user fecthed correctly -> compare passwords
+                    //use bcrypt.compare: first arg is the plainText password,, second arg is the crypted password stored in the db
+                    const isPasswordCorrect = await comparePassword(password, fecthedUser.password);
+
+                    if (!isPasswordCorrect){
+                        return res.status(HttpStatusCodes.UNAUTHORIZED).json({ error: "Error: invalid password!" });
+                    } else {
+                        //encrypt the new Password
+                        password = await hashPassword(newPassword);
+                    }
+
+                } else {
+                    return res.status(HttpStatusCodes.NOT_FOUND).json({ error: "Error: user not found!"});
+                }
+            } catch (error) {  
+                return res.status(HttpStatusCodes.INTERNAL_SERVER_ERROR).json({ error: "Error: it's not possible to get the user!" });
+            }
+        }
+        
+        
+        
+        
         
         try {
             const existentUser = new User(id, email, assignedName, familyName, password);
@@ -84,6 +121,10 @@ export function createUsersRouter({ userRepo }) {
         //convert id to number
         const id = parseInt(req.params.id);
 
+        if (isNaN(id)){ 
+            return res.status(HttpStatusCodes.BAD_REQUEST).json({ error: "Error: id is not a number!" });
+        }
+
         try {
             const user = await userRepo.getUserById(id);
             if (!user) {
@@ -104,6 +145,11 @@ export function createUsersRouter({ userRepo }) {
     router.delete("/:id", async (req, res) => {
         //convert id to number
         const id = parseInt(req.params.id);
+
+        if (isNaN(id)){
+            return res.status(HttpStatusCodes.BAD_REQUEST).json({ error: "Error: id is not a number!" });
+        }
+
         try {
             const user = await userRepo.deleteUser(id);
             if (!res){
